reverted next-auth and cleaned up auth page a tiny bit. more to do
