+++ src/react_frontend/package.json
@@ -23,7 +23,6 @@
         "eslint-config-next": "^13.0.7",
         "material-ui": "^0.20.2",
         "next": "13.0.5",
-        "next-auth": "^4.18.6",
         "next-stylus": "^0.0.2",
         "next-with-plugins": "^0.0.1",
         "react": "18.2.0",
