+++ src/react_frontend/src/services/api.ts
@@ -1,0 +1,7 @@
+const base_url = process.env.REACT_APP_BASE_URL || 'http://localhost:8000';
+
+const API = {
+  base_url
+}
+
+export default API
